Fix GCD import checkpoint offset type casting
- Added ::integer cast to jsonb_build_object offset parameter
- Fixes: ERROR could not determine data type of parameter $1

+++ backend/app/jobs/pipeline_scheduler.py
@@ -1175,7 +1175,7 @@ async def run_gcd_import_job(
                 # Store offset in state_data for resume
                 await db.execute(text("""
                     UPDATE pipeline_checkpoints
-                    SET state_data = jsonb_build_object('offset', :offset)
+                    SET state_data = jsonb_build_object('offset', :offset::integer)
                     WHERE job_name = :name
                 """), {"name": job_name, "offset": new_offset})
                 await db.commit()
